MVC2.0: Refactoring - remove dependency of MVC on Native API
