+++ fdbclient/IClientApi.h
@@ -148,6 +148,9 @@ public:
 	virtual Reference<IDatabase> createDatabase(const char* clusterFilePath) = 0;
 
 	virtual void addNetworkThreadCompletionHook(void (*hook)(void*), void* hookParameter) = 0;
+
+protected:
+	static IClientApi* localApi;
 };
 
 #endif
