+++ fdbclient/MultiVersionTransaction.actor.cpp
@@ -18,10 +18,9 @@
  * limitations under the License.
  */
 
-#include "fdbclient/CoordinationInterface.h"
 #include "fdbclient/MultiVersionTransaction.h"
 #include "fdbclient/MultiVersionAssignmentVars.h"
-#include "fdbclient/ThreadSafeTransaction.h"
+#include "fdbclient/ClientVersion.h"
 
 #include "flow/network.h"
 #include "flow/Platform.h"
@@ -29,6 +28,10 @@
 #include "flow/UnitTest.h"
 
 #include "flow/actorcompiler.h" // This must be the last #include.
+
+#ifdef FDBCLIENT_NATIVEAPI_ACTOR_H
+#error "MVC should not depend on the Native API"
+#endif
 
 void throwIfError(FdbCApi::fdb_error_t e) {
 	if (e) {
@@ -888,6 +891,30 @@ ACTOR Future<Void> timeoutImpl(Reference<ThreadSingleAssignmentVar<Void>> tsav, 
 	return Void();
 }
 
+namespace {
+
+void validateOptionValuePresent(Optional<StringRef> value) {
+	if (!value.present()) {
+		throw invalid_option_value();
+	}
+}
+
+int64_t extractIntOption(Optional<StringRef> value, int64_t minValue, int64_t maxValue) {
+	validateOptionValuePresent(value);
+	if (value.get().size() != 8) {
+		throw invalid_option_value();
+	}
+
+	int64_t passed = *((int64_t*)(value.get().begin()));
+	if (passed > maxValue || passed < minValue) {
+		throw invalid_option_value();
+	}
+
+	return passed;
+}
+
+} // namespace
+
 // Configure a timeout based on the options set for this transaction. This timeout only applies
 // if we don't have an underlying database object to connect with.
 void MultiVersionTransaction::setTimeout(Optional<StringRef> value) {
@@ -1467,7 +1494,7 @@ Reference<ClientInfo> MultiVersionApi::getLocalClient() {
 
 void MultiVersionApi::selectApiVersion(int apiVersion) {
 	if (!localClient) {
-		localClient = makeReference<ClientInfo>(ThreadSafeApi::api);
+		//		localClient = makeReference<ClientInfo>(ThreadSafeApi::api);
 	}
 
 	if (this->apiVersion != 0 && this->apiVersion != apiVersion) {
@@ -1482,6 +1509,8 @@ const char* MultiVersionApi::getClientVersion() {
 	return localClient->api->getClientVersion();
 }
 
+namespace {
+
 void validateOption(Optional<StringRef> value, bool canBePresent, bool canBeAbsent, bool canBeEmpty = true) {
 	ASSERT(canBePresent || canBeAbsent);
 
@@ -1492,6 +1521,8 @@ void validateOption(Optional<StringRef> value, bool canBePresent, bool canBeAbse
 		throw invalid_option_value();
 	}
 }
+
+} // namespace
 
 void MultiVersionApi::disableMultiVersionClientApi() {
 	MutexHolder holder(lock);
